引导模块第三次重构完成（但是还没有测试，增加了敏感操作认证、丰富前端websocket响应信息）

+++ src/main/java/edu/cqupt/mislab/erp/user/constant/UserConstant.java
@@ -10,7 +10,7 @@ public interface UserConstant {
     /*
     默认多少次登录失败后需要进行验证码校验
      */
-    Integer USER_MAX_WARN_TIMES_WITHOUT_VERIFICATION_CODE = 3;
+    Long USER_MAX_WARN_TIMES_WITHOUT_VERIFICATION_CODE = 3L;
 
     /*
     会话里面存储登录验证码的属性名称
